requirements.txt
Flask~=2.3.2
python-dotenv~=1.0.0
numpy~=1.24.3
flask-cors~=3.0.10
utm~=0.7.0
tqdm~=4.65.0